requirements.txt
appdirs==1.4.4
colored==1.4.2
cupy-cuda112==8.6.0
cycler==0.10.0
dataclasses==0.6
decorator==4.4.2
fastrlock==0.6
future==0.18.2
graphsurgeon @ file:///TensorRT-7.2.2.3/graphsurgeon/graphsurgeon-0.4.5-py2.py3-none-any.whl
kiwisolver==1.3.1
Mako==1.1.4
MarkupSafe==1.1.1
matplotlib==3.4.1
numpy==1.20.1
nvidia-pyindex==1.0.8
onnx==1.8.1
onnx-graphsurgeon==0.2.7
onnx-simplifier==0.3.4
onnx2pytorch==0.2.0
onnxoptimizer==0.2.5
onnxruntime==1.7.0
opencv-python==4.5.1.48
Pillow==8.1.1
polygraphy==0.25.0
protobuf==3.15.5
pycuda==2020.1
pyparsing==2.4.7
python-dateutil==2.8.1
pytools==2021.1
scipy==1.6.2
six==1.15.0
tensorrt @ file:///TensorRT-7.2.2.3/python/tensorrt-7.2.2.3-cp38-none-linux_x86_64.whl
torch==1.6.0 
torchsummary==1.5.1
torchvision 
typing-extensions==3.7.4.3
uff @ file:///TensorRT-7.2.2.3/uff/uff-0.6.9-py2.py3-none-any.whl
